pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin-autoreply"
version = "0.1.0"
description = "LinkedIn auto-reply bot"
requires-python = ">=3.11"
dependencies = [
  "scrapling[fetchers]>=0.4.7,<0.5",
  "pydantic>=2.6,<3",
  "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "ruff>=0.5",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["bot*"]
